Moved LegacySettings data type
